refactor: use sub folders inside components
