+++ src/App.tsx
@@ -1,8 +1,8 @@
 import "./App.css";
-import Shop from "./components/Shop";
-import Cart from "./components/Cart";
-import Error from "./components/Error";
-import Invoice from "./components/Invoice";
+import Shop from "./components/routes/Shop";
+import Cart from "./components/routes/Cart";
+import Error from "./components/routes/Error";
+import Invoice from "./components/routes/Invoice";
 import { Provider } from "./components/Context";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Nav from "./components/Nav";
